+++ packages/expo/src/provider/singleton/createClerkInstance.ts
@@ -21,7 +21,8 @@ import { MemoryTokenCache } from '../../cache/MemoryTokenCache';
 import type { TokenCache } from '../../cache/types';
 import { CLERK_CLIENT_JWT_KEY } from '../../constants';
 import { errorThrower } from '../../errorThrower';
-import { assertValidProxyUrl, isNative } from '../../utils';
+import { assertValidProxyUrl } from '../../utils/errors';
+import { isNative } from '../../utils/runtime';
 import type { BuildClerkOptions } from './types';
 
 /**