+++ packages/localizations/src/es-UY.ts
@@ -674,8 +674,8 @@ export const esUY: LocalizationResource = {
         messageLabel: 'Tu correo electrónico usa {{domain}}. ¿Querés agregarlo?',
       },
       formButtonPrimary__add: 'Agregar',
-      formFieldInputPlaceholder__domain: 'Escribí aquí tu dominio y hacé clic en Agregar para empezar',
-      formFieldLabel__domain: 'Dominios',
+      formFieldInputPlaceholder__domain: 'Agregar dominio',
+      formFieldLabel__domain: 'Dominio',
       removeDomainDialog: {
         cancelButton: undefined,
         removeButton: undefined,